Move tile business logic to operations folder
