+++ src/db/seed.ts
@@ -1,4 +1,4 @@
-import { tileActions } from '@/app/_actions/tile-actions'
+import { tileActions } from '@/operations/tile-actions'
 
 import { Location, Service } from './constants'
 import { client as dbClient } from './connection'
